+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-10T11:22:18.017Z",
+    "2026-03-10T12:22:30.846Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {