+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -403,7 +403,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KKBQR2N68S3PJBJNRC4BAV37"),
+  _id: z.string().default("01KKBV69Y5XNJW7BA5B01MZVAM"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,