+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-03-10T11:22:21.526Z",
+    "2026-03-10T12:22:34.163Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {