+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -152,7 +152,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KKBQR2NJ8XCMW9ARBR174QYX"),
+  _id: z.string().default("01KKBV69Y9KBS958AN22A25XKZ"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,